requirements.txt
requests==2.7.0
nltk==3.0.4
twitter==1.17.1
xmltodict==0.9.2
Pint==0.6
PyYAML==3.11